Change hashing algorithm
to be based on uniqueness of products of prime numbers.
C data types are not able to handle numbers these large so,
the code takes advantage of the fact that unsigned c numbers
wrap around upon overflowing.

Seems to be working.  An increase in tree height was predicted due to
less collisions but, the tree seems much taller now.  I am not sure
if it is a bug because, anagrams could not be found if it was not
colliding at appropriate times.  Will look into further.

+++ treapset.h
@@ -2,7 +2,7 @@
 #define TREAPSET_H
 
 /*change to fit your program*/
-typedef int TREAPTYPE;
+typedef unsigned long TREAPTYPE;
 
 typedef struct treap {
 	int priority;
